+++ install.py
@@ -24,7 +24,7 @@ class MesowxInstaller(ExtensionInstaller):
         raw_key = random_password()
 
         super(MesowxInstaller, self).__init__(
-            version='0.5.0',
+            version='0.5.1',
             name='mesowx',
             description='Weather station console web front-end with'
                         'real-time updating and dynamic charting.',