+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 * Combine the 3 scripts into the one - mesowx.py
 * Works with weewx4 and python 3, also with python2.7
 * Use the wee_extension installer to perform a (mostly) hands off install of the local setup,
-    and minimize interaction required for the remote install.
+    and minimize the interaction required for the remote install.
 * Adjust the script, add tmpl's, use randompassword generator for security key insertion.
 * Remote install works with archive and raw, Local works with raw database and weewx database,
     RetainLoopValues remains untested.